requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
ultralytics==8.0.196
Werkzeug==2.3.7
